requirements.txt
django==4.2.5
django-environ==0.11.2
django-storages==1.14.1
djangorestframework==3.14.0
psycopg==3.1.12
zappa==0.58.0